+++ changelog.md
@@ -3,11 +3,12 @@
 
 
 ## Standard library additions and changes
-- Added `args` optional param to `osproc.execCmdEx`
 
 - `prelude` now works with the JavaScript target.
 
 - Added `ioutils` module containing `duplicate` and `duplicateTo` to duplicate `FileHandle` using C function `dup` and `dup2`.
+
+- Added `osproc.execArgs`
 
 
 ## Language changes